✨[Feat] 패키지명 입력 후 완료 API 구현
