+++ src/main/java/com/eventory/server/global/apipayload/exception/handler/ProductHandler.java
@@ -1,0 +1,11 @@
+package com.eventory.server.global.apipayload.exception.handler;
+
+import com.eventory.server.global.apipayload.code.BaseErrorCode;
+import com.eventory.server.global.apipayload.exception.GeneralException;
+
+public class ProductHandler extends GeneralException {
+
+    public ProductHandler(BaseErrorCode errorCode) {
+        super(errorCode);
+    }
+}
